requirement.txt
aiohttp==3.8.1
aiosignal==1.2.0
alpha-vantage==2.3.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
colorama==0.4.5
DateTime==4.7
debugpy==1.6.2
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.4
dj-database-url==1.0.0
Django==4.0.6
django-admin-logs==1.0.2
django-ckeditor==6.5.1
django-email-verification==0.3.1
django-extensions==3.2.0
django-heroku==0.3.1
django-js-asset==2.0.0
django-rest-framework==0.1.0
Django-Verify-Email==2.0.3
djangorestframework==3.13.1
entrypoints==0.4
executing==0.9.1
fastjsonschema==2.16.1
filelock==3.7.1
forex-python==1.8
frozenlist==1.3.1
gevent==21.12.0
greenlet==1.1.3
gunicorn==20.1.0
idna==3.3
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.9.1
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.2
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.2
plotly==5.10.0
ply==3.11
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2==2.9.3
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
Pyomo==6.4.2
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.2.1
pywin32==304
pywinpty==2.0.6
pyzmq==23.2.0
qrcode==7.3.1
qtconsole==5.3.1
QtPy==2.1.0
requests==2.28.1
requests-toolbelt==0.9.1
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
StackAPI==0.2.0
tenacity==8.0.1
terminado==0.15.0
tinycss2==1.1.1
tornado==6.2
traitlets==5.3.0
twilio==7.13.0
tzdata==2022.1
urllib3==1.26.12
validate-email==1.3
virtualenv==20.15.1
wcwidth==0.2.5
webencodings==0.5.1
websocket==0.2.1
whitenoise==6.2.0
widgetsnbextension==3.6.1
yarl==1.8.1
zope.event==4.5.0
zope.interface==5.4.0
